Added support to write to write log entries to DB
With this check-in, the log entries are written to the DB, but the get
method to recover the data have not been written, yet. So you need to
look at them in phpMyAdmin, for the time being. Eventually, (soon), the
log queries will be read as get requests from the logger.php URI.